performance typeScript improve project structure
