+++ package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "instead": "^1.0.3",
+    "prop-types": "^15.8.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-router-dom": "^7.1.3"
